Minor cleanup and major fix for using closures with give().
Closures with give() were designed so you could modify the jest via the mime,
however, due to the fact that they were run early, the expectations for various
calls got in the wrong order.  This has been resolved, as the closures are now not
run until the actual method is called on the jest.