Fixed class_exists() triggering autoload.  Added ability to extend existing class.
You can now use Mime to extend a class all the way down to an original
existing class that is not also mimed.  HOWEVER!  It should be noted that at
this point you cannot currently make it mimick anything, it essentially just
becomes an alias.  This is only really useful at the moment for creating
Exceptions (assuming your exceptions are pretty normal).